### 2019-03-19 #1334 @whedon
[PRE REVIEW]: GUESSmyLT: Software to guess the RNA-Seq library type of paired and single end read files
**Submitting author:** @Juke34 (<a href="http://orcid.org/0000-0002-6629-0173">Jacques Dainat</a>)
**Repository:** <a href="https://github.com/NBISweden/GUESSmyLT" target ="_blank">https://github.com/NBISweden/GUESSmyLT</a>
**Version:** v1.0

Thanks for submitting your paper to JOSS @Juke34. Currently, there isn't an JOSS editor assigned to your paper.
