### 2019-03-19 #1334 @arfon
@Juke34 - what you're trying to do with the `*` affiliation won't work sorry. The affiliation indexes need to be numerical.
