### 2019-03-19 #1334 @Juke34
Is there any advised way to pinpoint equal contribution ?
