### 2019-03-19 #1334 @pjotrp
I would add it as a footnote. You can even point to the github contributions graph.
