### 2019-03-20 8b60f51
remove star from affiliation
