+++ paper.bib
@@ -1,6 +1,5 @@
 @article{RSeQC,
-    Adsurl = {http://doi.org/10.1093/bioinformatics/bts356},
-    Author = {Wang, L., Wang, S., & Li, W.},
+    Author = {Wang L., Wang S., & Li W.},
     Journal = {Bioinformatics},
     Keywords = {RSeQC},
     Title = {{RSeQC: quality control of RNA-seq experiments}},
@@ -10,7 +9,7 @@
     doi = {10.1093/bioinformatics/bts356}
 }
 @article{Salmon,
-    Author = {Patro, R., Duggal, G., Love, M. I., Irizarry, R. A., & Kingsford, C.},
+    Author = {Patro R., Duggal G., Love M. I., Irizarry R. A., & Kingsford C.},
     Journal = {Nature Methods},
     Keywords = {Salmon},
     Title = {{Salmon provides fast and bias-aware quantification of transcript expression}},
