### 2019-03-20 #1334 @Juke34
I had apparently forgot a sentence. I also fixed a typo and added a figure. I'm not sure having proceeding properly with the figure. Let's see the new pdf.
