### 2019-03-22 #1344 @whedon
[REVIEW]: GUESSmyLT: Software to guess the RNA-Seq library type of paired and single end read files
**Submitting author:** @Juke34 (<a href="http://orcid.org/0000-0002-6629-0173">Jacques Dainat</a>)
**Repository:** <a href="https://github.com/NBISweden/GUESSmyLT" target ="_blank">https://github.com/NBISweden/GUESSmyLT</a>
**Version:** v1.0
