### 2019-03-22 #1344 @lpantano
Thanks @mistrm82 to take time to review this paper!

@Juke34, please, go over the items that the reviewer will need to check and make sure all are good to go.

I think @mistrm82 will start reviewing the paper around Next Thursday.

Cheers
