### 2019-03-22 c668b25
add  Community guidelines + move paper in dedicated folder
