### 2019-03-22 #1344 @Juke34
Everything sounds ok except we didn't have any Community guidelines. Now is done. I have re-flagged this last version as v1.0. I took the opportunity to add acknowledgements in the paper too, and move all the paper related files into the paper dedicated folder. So, I guess we should compile the pdf again...
