### 2019-04-10 #1344 @mistrm82
@Juke34 I am having some problems installing (created an issue) so I have not been able to assess the "Functionality" section
