### 2019-04-11 #1344 @mistrm82
I think this is a great tool with functionality that meets a need for researchers doing NGS analysis. The problems I encountered is that installation is quite involved. I created a conda environment with the prerequisites (biopython, pysam, bcbio-gff) which was not too much of an issue. However, when trying to run some commands I encountered errors due to the lack of other required programs (i.e busco, trinity). Working on a cluster environment, the installation of these tools is not always straightforward. Is there any way to have these extra programs be installed as dependencies when you run the setup.py install? This would make it easier. Since I couldn't do these installations I was unable to fully test out functionality.
