### 2019-04-12 c44bcc8
Update README.md
 fix #10 - Wrong statements removed from the readme.
