### 2019-04-18 #1344 @lpantano
Hi @Juke34, Do you think you can address @mistrm82 comment to get a better way to get the tool installed?
