### 2019-04-19 #1344 @Juke34
Yes I think we can make a bioconda package.
