### 2019-05-14 0cd71d2
separate samtools to bowtie fixed. + whole path for gzip fixed
