### 2019-05-14 6a96ff2
add infer_lib.py as script
