### 2019-05-14 4396551
move around infer_lib.py to have use it non-executable for portability. Call it with python prefix in run,shell
