### 2019-06-04 1a3e8c8
add busco data within the package. Untar during usage
