### 2019-06-09 #1344 @Juke34
Thank you for your patience. Here are the updates: We have slightly modified the installation. We have created a PyPi package (https://pypi.org/project/guessmylt/), allowing installation using pip.  So, using `pip install GUESSmyLT` will install most of the dependencies. Still Bowtie2, BUSCO and Trinity (depending of the user need Trinity is not mandatory) has to be installed aside.

Working on conda to package the whole, I have seen differences in the BUSCO results between osx and linux, and Trinity does not work on osx. So I have created a Bioconda recipe that should be compiled on linux (https://github.com/bioconda/bioconda-recipes/pull/15786). Each package added to Bioconda also has a corresponding Docker BioContainer automatically created and uploaded to Quay.io. Once the container available, the reproducibility problem (different result between mac and linux) would be solved. The thing is, my Bioconda package is currently stuck at the check step because the bioconda-utils is encountered some issues (https://github.com/bioconda/bioconda-utils/pull/567) that have to be resolved first.

May the simplification with `pip install GUESSmyLT` enough for publication?
