### 2019-06-10 #1344 @epruesse
@Juke34 The problem with your package is not related to the bioconda-utils PR mentioned.

- You are currently installing files to `site-packages/example_run` which should go to `site-packages/GUESSmlLT` or be omitted from distributed package.
- The package currently fails in testing. If you look at the build logs, you will find at the bottom during the tests in an isolated minimal container an exception raised by gitpython indicating that `git` is missing. It's likely due to a missing run dependency on `git` itself.
