### 2019-06-11 b214cbf
move test in GUESSmyLT package
