+++ MANIFEST.in
@@ -1,15 +1,3 @@
-include GUESSmyLT/bowtie2
-include GUESSmyLT/busco
-include GUESSmyLT/inferr
-include GUESSmyLT/Snakefile
-include GUESSmyLT/trinity
-include GUESSmyLT/inferr_lib.py
-include GUESSmyLT/check_programs
-include GUESSmyLT/subsample
-include GUESSmyLT/samtools
-include example_run/example.py
-include example_run/4_r1.fastq.gz
-include example_run/4_r2.fastq.gz
-include example_run/4.fa
-include GUESSmyLT/data/eukaryota_odb9.tar.gz
-include GUESSmyLT/data/bacteria_odb9.tar.gz
+include GUESSmyLT/*
+include LICENSE
+include README.md
