### 2019-06-11 988499b
add LICENSE and simplify
