### 2019-06-12 70670a4
set AUGUSTUS_CONFIG_PATH automatically if missing while using BUSCO
