### 2019-06-13 f93d6dc
add relative orientatio of reads. add info when no genes to predict the lib. Change results print. Remove extra loop over reads not needed
