+++ GUESSmyLT/inferr_lib.py
@@ -110,7 +110,7 @@ def extract_genes(annotation, run_name):
     print("Number of genes extracted: %d\n" % (len(correct_genes.features)))
     return correct_genes
 
-def infer_paired_region(genes):
+def infer_paired_region_with(genes):
     '''
     Function for inferring paired library-type by looking at a regions corresponding to genes
     '''
@@ -131,53 +131,95 @@ def infer_paired_region(genes):
         start = int(gene.location.start)
         stop = int(gene.location.end)
         strand = gene.strand
-        reads = []
-       
+
         # Get reads mapped to a specific contig and in a sequence range
         # TODO: Look into optimizing this step, only take a subset (1000ish) reads? samfile.mate is not made for high throughput
         try:
             for read in samfile.fetch(contig, start, stop):
                 if not read.mate_is_unmapped and read.is_read1:
-                    reads.append([read, samfile.mate(read)])
+                    first = read
+                    second = samfile.mate(read)
+                    try:
+                        lib = ''
+                        if not first.is_reverse:
+                            lib += 'f'
+                        else:
+                            lib += 'r'
+                        if not second.is_reverse:
+                            lib += 'f'
+                        else:
+                            lib += 'r'
+                        # Gene on sense strand
+                        if strand == 1:
+                            if first.reference_start > second.reference_start:
+                                # Flip order of reads
+                                lib = lib[::-1]
+                                lib += '_first'
+                            elif first.reference_start < second.reference_start:
+                                lib += '_second'
+                            else:
+                                lib = 'undecided'
+                        # Gene on antisense
+                        elif strand == -1:
+                            if first.reference_start > second.reference_start:
+                                # Flip order of reads
+                                lib = lib[::-1]
+                                lib += '_second'
+                            elif first.reference_start < second.reference_start:
+                                lib += '_first'
+                            else:
+                                lib = 'undecided'
+                        libs[lib] += 1
+                    except: libs['undecided'] +=1 #Some reads missing start or end-values
         except ValueError as message:
-            print(message)
+            print(message)         
 
-        # Check lib-type of reads
-        for read in reads:
-            first = read[0]
-            second = read[1]
-            try:
-                lib = ''
-                if not first.is_reverse:
-                    lib += 'f'
-                else:
-                    lib += 'r'
-                if not second.is_reverse:
-                    lib += 'f'
-                else:
-                    lib += 'r'
-                # Gene on sense strand
-                if strand == 1:
+    return libs
+
+def infer_paired_region_without():
+    '''
+    Function for inferring paired library-type by looking at a regions corresponding to genes
+    '''
+    # Counters for the different lib-types
+    libs = {
+        'fr': 0,
+        'rf': 0,
+        'ff': 0,
+        'undecided0': 0
+    }
+
+    # Get reads 
+    # TODO: Look into optimizing this step, only take a subset (10000ish) reads? samfile.mate is not made for high throughput
+    try:
+        max_read=20000
+        count=0
+        print ("As no gene was found we focus only on "+str(max_read)+" first reads (both reads mapped)")
[…]
