### 2019-06-20 #1344 @lpantano
I have added the right tag to get this merged, if not I could ping the bioconda team. Thanks for the effort.
