### 2019-06-24 #1344 @mistrm82
Thanks @Juke34  for both the PyPi package (https://pypi.org/project/guessmylt/) and the bioconda recipe. The problem I had previously was the inability to get Busco and so I think I will hold off on testing the PyPi  package and wait for the PR merge from bioconda
