### 2019-06-26 #1344 @lpantano
@mistrm82, the package should be able through bioconda today, so feel free to test it from tomorrow. Let us know if you find any issue. Thanks!
