### 2019-07-03 #1344 @mistrm82
Sorry for the delay. I was successfully able to install via bioconda and use the tool. Thanks @Juke34 everything looks great to me
