### 2019-07-03 #1344 @lpantano
@Juke34, We are almost there. Can you generate a Zenodo archive pointing to your github current version and making sure title and authors match the Zenodo link? Thanks
