### 2019-07-04 #1344 @Juke34
Here is the dog: 10.5281/zenodo.3268394
