### 2019-07-05 #1344 @lpantano
Thanks for the hard work to the authors and reviewers. @openjournals/joss-eics this is accepted now.
