### 2019-07-05 #1344 @labarba
I'm confused by the author list: are commas missing, or something? It looks like some authors have a single-letter last name. Are first and last names being displayed in the wrong order?
