### 2019-07-07 #1344 @whedon
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
